Fixes: - Styles updates to meet PR comments standards (responsive fixed) - Repository is now named Dummy. Other provisional renaming issues have been fixed.
