+++ server/src/main/java/com/knowy/server/service/OptionQuizService.java
@@ -1,8 +1,8 @@
 package com.knowy.server.service;
 
 
-import com.knowy.server.controller.model.OptionQuizDTO;
-import com.knowy.server.repository.OptionsQuizRepository;
+import com.knowy.server.controller.dto.OptionQuizDTO;
+import com.knowy.server.repository.OptionsQuizRepositoryDummy;
 import org.springframework.stereotype.Service;
 
 
@@ -11,15 +11,15 @@ import java.util.List;
 
 @Service
 public class OptionQuizService {
-    private final OptionsQuizRepository optionsQuizRepository;
+    private final OptionsQuizRepositoryDummy optionsQuizRepositoryDummy;
 
 
-    public OptionQuizService(OptionsQuizRepository optionsQuizRepository) {
-        this.optionsQuizRepository = optionsQuizRepository;
+    public OptionQuizService(OptionsQuizRepositoryDummy optionsQuizRepositoryDummy) {
+        this.optionsQuizRepositoryDummy = optionsQuizRepositoryDummy;
     }
 
 
     public List<OptionQuizDTO> getOptionsForQuiz(int quizID){
-        return optionsQuizRepository.findOptionsByQuizId(quizID);
+        return optionsQuizRepositoryDummy.findOptionsByQuizId(quizID);
     }
 }
