+++ server/src/main/resources/static/styles/components/cards-questionnaire/options-quiz.scss
@@ -8,7 +8,7 @@
 	width: 80%;
 
 	@media (max-width: 768px) {
-		width: 100%;
+		width: 80%;
 		padding: 0 1rem;
 	}
 }
