+++ server/src/main/resources/static/styles/components/cards-questionnaire/questionComponent.scss
@@ -2,7 +2,6 @@
 
 .question-component-container {
 
-	margin:1em;
 	width: 80%;
 	display:flex;
 	flex-direction: column;
