--- server/src/main/resources/static/styles/pages/testOptionsQuiz.scss
@@ -1,3 +1,0 @@
-@use '../global-vars' as *;
-@use '../components/navbar';
-@use '../layouts/quiz-layout';
