+++ server/src/main/resources/static/styles/pages/tests.scss
@@ -1,0 +1,83 @@
+@use '../global-vars' as *;
+@use '../components/navbar';
+@use '../layouts/quiz-layout';
+
+
+body{
+	font-family:excalifont;
+}
+
+.test-container {
+	max-width: 800px;
+	margin: 2rem auto;
+	padding: 1.5rem;
+	background-color: $primary-300;
+	border-radius: 20px;
+}
+
+.answers-container{
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	margin:1em;
+}
+.question-container{
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	margin:1em;
+}
+
+.answer-button{
+	text-align:center;
+
+	.answer-button__element {
+		width: 80%;
+		background-color: $primary-500;
+		color: white;
+		border: none;
+		padding: 0.8rem 1.5rem;
+		font-size: 1rem;
+		font-weight: bold;
+		border-radius: 15px;
+		cursor: pointer;
+		transition: background-color 0.3s;
+
+		&:hover {
+			background-color: $primary-700;
+		}
+	}
+}
+
+.navigation-buttons {
+	width:80%;
+	display: flex;
+	align-self:center;
+	justify-content: space-between;
+	margin: 1em auto 1em auto;
+
+	@media (max-width: 576px) {
+		flex-direction: row;
+		gap: 10px;
+		padding:0 10px;
+		align-items: center;
+		width:85%;
+	}
+
+	.navigation-buttons__element {
+		background-color: $primary-500;
+		color: white;
+		border: none;
+		padding: 0.8rem 1.5rem;
+		font-size: 1rem;
+		font-weight: bold;
+		border-radius: 15px;
+		cursor: pointer;
+		transition: background-color 0.3s;
+
+
+		&:hover {
+			background-color: $primary-700;
+		}
+	}
+}
