+++ server/src/main/resources/templates/components/cards-quiz/quiz-header.html
@@ -1,14 +1,15 @@
 <div th:fragment="quiz-header">
-	<div class="curso-banner" th:object="${quizLayout}">
-		<h2 th:text="*{courseName}">Angular 19 nivel intermedio + Firebase</h2>
-	</div>
+	<div class="curso-header">
+		<div class="curso-banner" th:object="${quizLayout}">
+			<h2 th:text="*{courseName}">Angular 19 nivel intermedio + Firebase</h2>
+		</div>
 
-
-	<!-- Datos de navegación -->
-	<div class="info-panel" th:object="${quizLayout}">
-		<div class="info-panel__element">Lección: <span class="info-panel__element__content" th:text="*{lessonNumber}">0</span></div>
-		<div class="info-panel__element">Nº Tarjeta: <span class="info-panel__element__content" th:text="*{quizCardNumber}">0</span></div>
-		<div class="info-panel__element">Racha: <span class="info-panel__element__content" th:text="*{cardStreak}">0</span></div>
-		<div class="info-panel__element">Precisión: <span class="info-panel__element__content" th:text="*{correctPercentage} + '%'">0%</span></div>
+		<!-- Datos de navegación -->
+		<div class="info-panel" th:object="${quizLayout}">
+			<div class="info-panel__element">Lección: <span class="info-panel__element__content" th:text="*{lessonNumber}">0</span></div>
+			<div class="info-panel__element">Nº Tarjeta: <span class="info-panel__element__content" th:text="*{quizCardNumber}">0</span></div>
+			<div class="info-panel__element">Racha: <span class="info-panel__element__content" th:text="*{cardStreak}">0</span></div>
+			<div class="info-panel__element">Precisión: <span class="info-panel__element__content" th:text="*{correctPercentage} + '%'">0%</span></div>
+		</div>
 	</div>
 </div>
