--- server/src/main/resources/templates/pages/testOptionsQuiz.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="es"
-	  layout:decorate="~{layouts/user-postlogin-layout}"
-	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-	  xmlns:th="http://www.thymeleaf.org">
-<head>
-	<title>Explicación de la lección</title>
-	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css">
-
-</head>
-<body >
-<!-- I use this page ONLY for testing the quiz-options. It's not necessary. In case that you want, delete this page. -->
-<main layout:fragment="main-content" class="main-content">
-	<div th:replace="~{components/cards-quiz/options-quiz :: options-quiz}"></div>
-</main>
-
-</body>
-</html>
