+++ server/src/main/resources/templates/pages/tests.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="es"
+	  layout:decorate="~{layouts/quiz-layout.html}"
+	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+	  xmlns:th="http://www.thymeleaf.org">
+<head>
+	<title>Test</title>
+	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css">
+	<link rel="stylesheet" th:href="@{/styles/pages/tests.css}">
+</head>
+<body >
+<!-- I use this page ONLY for testing the quiz-options. It's not necessary. In case that you want, delete this page. -->
+<main layout:fragment="main-content">
+	<!--CARD LAYOUT HEADER-->
+	<div th:replace="~{components/cards-quiz/quiz-header :: quiz-header}"></div>
+
+	<!--PROBLEM DEFINITION-->
+	<div class="question-container">
+		<div th:replace="~{components/cards-quiz/question-component :: question-component}"></div>
+	</div>
+
+	<!--POSSIBLE ANSWERS COMPONENT-->
+	<div class="answers-container">
+		<div th:replace="~{components/cards-quiz/options-quiz :: options-quiz}"></div>
+	</div>
+
+	<!--NAVIGATION BUTTONS-->
+	<div layout:fragment="quiz-navigation-buttons">
+		<div class="answer-button">
+			<button type="submit" class="answer-button__element">Responder</button>
+		</div>
+
+		<div class="navigation-buttons">
+			<a href="#">
+				<button class="navigation-buttons__element">← Anterior</button>
+			</a>
+
+			<a href="#">
+				<button class="navigation-buttons__element">Siguiente →</button>
+			</a>
+
+		</div>
+	</div>
+</main>
+</body>
+</html>
